requirements.txt
aiofiles==0.7.0
aioice==0.7.5
aiortc==1.2.1
av==8.0.3
cffi==1.14.6
cryptography==3.4.8
dnspython==2.1.0
google-crc32c==1.1.5
httptools==0.3.0
multidict==5.1.0
netifaces==0.11.0
numpy==1.21.2
Pillow==8.3.2
pycparser==2.20
pyee==8.2.2
pylibsrtp==0.6.8
sanic==21.6.2
sanic-routing==0.7.1
ujson==4.1.0
uvloop==0.16.0
websockets==9.1
